refactor:#265 client meal post 리팩토링
