+++ src/app/(client)/meal/post/_components/add-meal-drawer.tsx
@@ -18,9 +18,9 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { createAiRequestByText, createAiResponse } from '@/apis/analysis-request.api';
 import { useRouter } from 'next/navigation';
 import Modal from '@/components/commons/modal';
-import { ERROR_MESSAGES } from '../constants/analysis-error.constant';
-import { FoodFormValues, formSchema } from '../schemas/add-meal.schema';
-import { revalidate } from '../utils/revalidate';
+import { ERROR_MESSAGES } from '../_constants/analysis-error.constant';
+import { FoodFormValues, formSchema } from '../_schemas/add-meal.schema';
+import { revalidate } from '../_utils/revalidate.util';
 
 const AddMealDrawer = () => {
   const [isAnalyzing, setIsAnalyzing] = useState<boolean>(false);
